requirements.txt
boto3
pymongo
python-dotenv
sagemaker
